cloud_functions/index_metadata/requirements.txt
# Cloud Functions Requirements (v4.0.0)
# Python 3.11 runtime

# Cloud Functions Framework
functions-framework==3.*

# Google Cloud Libraries
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.1
google-cloud-aiplatform==1.38.1
